requirements.txt
djangorestframework
Django
PyJWT
djangorestframework-simplejwt
